requirements.txt
opencv-python
numpy
tensorflow
Pillow
scikit-learn
